# docs/swagger.yaml
definitions:
  ds.Meals:
    properties:
      id:
        type: integer
      image_url:
        type: string
      meal_brand:
        type: string
      meal_detail:
        type: string
      meal_info:
        type: string
      meal_weight:
        type: string
      status:
        type: boolean
    type: object
...
  schemas.AddMealToMilkReqResponse:
    properties:
      mealID:
        type: integer
      messageResponse:
        type: string
      milkRequestID:
        type: integer
    type: object
...
  schemas.CreateMealRequest:
    properties:
      image_url:
        type: string
      meal_brand:
        type: string
      meal_detail:
        type: string
      meal_info:
        type: string
      meal_weight:
        type: string
      status:
        type: boolean
    type: object
  schemas.CreateMealResponse:
    properties:
      id:
        type: integer
      messageResponse:
        type: string
    type: object
...
  schemas.DeleteMealResponse:
    properties:
      id:
        type: integer
      messageResponse:
        type: string
    type: object
...
  schemas.GetAllMealsResponse:
    properties:
      count:
        type: integer
      count_meals_in_draft_request:
        type: integer
      meals:
        items:
          $ref: '#/definitions/ds.Meals'
        type: array
      milk_req_ID:
        type: integer
    type: object
...
  schemas.GetMealResponse:
    properties:
      meal:
        $ref: '#/definitions/ds.Meals'
    type: object
...
  schemas.ResponseMessage:
    type: object
...
  schemas.UpdateMealRequest:
    properties:
      image_url:
        type: string
      meal_brand:
        type: string
      meal_detail:
        type: string
      meal_info:
        type: string
      meal_weight:
        type: string
      status:
        type: boolean
    type: object
  schemas.UpdateMealResponse:
    properties:
      id:
        type: integer
      messageResponse:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/meal:
    post:
      consumes:
      - application/json
      description: Create meal with properties
      parameters:
      - description: Meal data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateMealRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.CreateMealResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Create meal
      tags:
      - meals
  /api/meal/{ID}:
    delete:
      consumes:
      - application/json
      description: Delete meal using it's ID
      parameters:
      - description: Meal ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.DeleteMealResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Delete meal by ID
      tags:
      - meals
    get:
      consumes:
      - application/json
      description: Get info about meal using its ID
      parameters:
      - description: Meal ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetMealResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Get meal by ID
      tags:
      - meals
    put:
      consumes:
      - application/json
      description: Update meal using it's ID with parametres
      parameters:
      - description: Meal ID
        in: path
        name: ID
        required: true
        type: string
      - description: Update meal data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateMealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateMealResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Update meal by ID
      tags:
      - meals
  /api/meal/change_pic/{ID}:
    post:
      consumes:
      - application/json
      description: Delete meal using it's ID
      parameters:
      - description: Meal ID
        in: path
        name: ID
        required: true
        type: string
      - description: File
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Picture was changed sucessfully
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Change picture By ID
      tags:
      - meals
  /api/meal_to_milk_request/{ID}:
    post:
      consumes:
      - application/json
      description: This endpoint allows you to add a meal to a milk request by it's
        ID.
      parameters:
      - description: Meal ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meal added successfully
          schema:
            $ref: '#/definitions/schemas.AddMealToMilkReqResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Add meal to milk request
      tags:
      - meals
  /api/meals:
    get:
      consumes:
      - application/json
      description: Returns a list of all meals.
      produces:
      - application/json
      responses:
        "200":
          description: List of meals retrieved successfully
          schema:
            $ref: '#/definitions/schemas.GetAllMealsResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Get all meals
      tags:
      - meals
...
swagger: "2.0"